+++ src/usaxs/plans/plans_tune.py
@@ -1,0 +1,537 @@
+"""
+user-facing scans
+"""
+
+import datetime
+import logging
+import os
+import time
+from collections import OrderedDict
+from typing import Any
+from typing import Dict
+from typing import Optional
+
+from .mono_feedback import MONO_FEEDBACK_OFF
+from .mono_feedback import MONO_FEEDBACK_ON
+from apsbits.core.instrument_init import oregistry
+from apstools.devices import SCALER_AUTOCOUNT_MODE
+from apstools.plans import restorable_stage_sigs
+from bluesky import plan_stubs as bps
+from bluesky import preprocessors as bpp
+from .axis_tuning import tune_mr
+from .axis_tuning import tune_ar
+from .axis_tuning import tune_a2rp
+from usaxs.startup import suspend_BeamInHutch
+from usaxs.startup import suspend_FE_shutter
+from usaxs.utils.emails import email_notices
+from usaxs.utils.override_parameters import user_override
+from .requested_stop import IfRequestedStopBeforeNextScan
+from .mode_changes import mode_USAXS
+from ..utils.a2q_q2a import q2angle
+from .command_list import after_plan
+from .command_list import before_plan
+from .mode_changes import mode_USAXS
+from .mode_changes import mode_SAXS
+from .mode_changes import mode_WAXS
+from .requested_stop import IfRequestedStopBeforeNextScan 
+from .sample_transmission import measure_SAXS_Transmission
+from .sample_transmission import measure_USAXS_Transmission
+from .filter_changes import insertSaxsFilters
+from .filter_changes import insertWaxsFilters
+from .area_detector import areaDetectorAcquire
+from .amplifiers import autoscale_amplifiers
+from .I0_controls import I0_controls
+from .I00_controls import I00_controls
+from .upd_controls import upd_controls
+from .trd_controls import trd_controls
+from .sample_imaging import record_sample_image_on_demand
+
+
+# from usaxs.utils.setup_new_user import cleanupText
+# from usaxs.utils.setup_new_user import techniqueSubdirectory
+# from usaxs.utils.user_sample_title import getSampleTitle
+
+# # Constants
+# tune_m2rp = oregistry["tune_m2rp"]
+# tune_ar = oregistry["tune_ar"]
+# tune_a2rp = oregistry["tune_a2rp"]
+# NOTIFY_ON_BADTUNE = oregistry["NOTIFY_ON_BADTUNE"]
+
+# tune_mr = oregistry["tune_mr"]
+# uascan = oregistry["uascan"]
+# NOTIFY_ON_BAD_FLY_SCAN = oregistry["NOTIFY_ON_BAD_FLY_SCAN"]
+
+logger = logging.getLogger(__name__)
+
+
+# # these two templates match each other, sort of
+# AD_FILE_TEMPLATE = "%s%s_%4.4d.hdf"
+# LOCAL_FILE_TEMPLATE = "%s_%04d.hdf"
+MASTER_TIMEOUT = 60
+# user_override.register("useDynamicTime")
+
+# # Make sure these are not staged. For acquire_time,
+# # any change > 0.001 s takes ~0.5 s for Pilatus to complete!
+# DO_NOT_STAGE_THESE_KEYS___THEY_ARE_SET_IN_EPICS = """
+#     acquire_time acquire_period num_images num_exposures
+# """.split()
+
+# # Device and plan instances from oregistry (allowed list)
+mono_shutter = oregistry["mono_shutter"]
+usaxs_shutter = oregistry["usaxs_shutter"]
+# usaxs_shutter = oregistry["usaxs_shutter"]
+# ar_start = oregistry["ar_start"]
+guard_slit = oregistry["guard_slit"]
+# lax_autosave = oregistry["lax_autosave"]
+# m_stage = oregistry["m_stage"]
+monochromator = oregistry["monochromator"]
+terms = oregistry["terms"]
+s_stage = oregistry["s_stage"]
+# saxs_det = oregistry["saxs_det"]
+# saxs_stage = oregistry["saxs_stage"]
+# struck = oregistry["struck"]
+# terms = oregistry["terms"]
+# usaxs_flyscan = oregistry["usaxs_flyscan"]
+# usaxs_q_calc = oregistry["usaxs_q_calc"]
+usaxs_slit = oregistry["usaxs_slit"]
+user_data = oregistry["user_device"]
+scaler0 = oregistry["scaler0"]
+# waxs_det = oregistry["waxs_det"]
+m_stage = oregistry["m_stage"]
+a_stage = oregistry["a_stage"]
+d_stage = oregistry["d_stage"]
+# flyscan_trajectories = oregistry["flyscan_trajectories"]
+# # Plan helpers (if available in oregistry)
+# mode_USAXS = oregistry["mode_USAXS"]
+# mode_SAXS = oregistry["mode_SAXS"]
+# mode_WAXS = oregistry["mode_WAXS"]
+# record_sample_image_on_demand = oregistry["record_sample_image_on_demand"]
+# measure_USAXS_Transmission = oregistry["measure_USAXS_Transmission"]
+# measure_SAXS_Transmission = oregistry["measure_SAXS_Transmission"]
+# insertSaxsFilters = oregistry["insertSaxsFilters"]
+# insertWaxsFilters = oregistry["insertWaxsFilters"]
+# areaDetectorAcquire = oregistry["areaDetectorAcquire"]
+# autoscale_amplifiers = oregistry["autoscale_amplifiers"]
+# I0_controls = oregistry["I0_controls"]
+# I00_controls = oregistry["I00_controls"]
+# upd_controls = oregistry["upd_controls"]
+# trd_controls = oregistry["trd_controls"]
+# scaler0 = oregistry["scaler0"]
+# scaler1 = oregistry["scaler1"]
+# constants = oregistry["constants"]
+ 
+
+
+
+
+
+@bpp.suspend_decorator(suspend_FE_shutter)
+@bpp.suspend_decorator(suspend_BeamInHutch)
+def preUSAXStune(md={}
+):
+    """
+    Tune the USAXS optics in any mode, is safe.
+
+    Parameters
+    ----------
+    md : Optional[Dict[str, Any]], optional
+        Metadata dictionary, by default None
+    RE : Optional[Any], optional
+        Bluesky RunEngine instance, by default None
+    bec : Optional[Any], optional
+        Bluesky Live Callbacks instance, by default None
+    specwriter : Optional[Any], optional
+        SPEC file writer instance, by default None
+
+    Returns
+    -------
+    Generator[Any, None, None]
+        A generator that yields plan steps
+
+    USAGE:  ``RE(preUSAXStune())``
+    """
+
+    yield from bps.mv(
+        monochromator.feedback.on,
+        MONO_FEEDBACK_ON,
+        mono_shutter,
+        "open",
+        usaxs_shutter,
+        "close",
+        timeout=MASTER_TIMEOUT,
+    )
+    yield from IfRequestedStopBeforeNextScan()  # stop if user chose to do so.
+
+    yield from mode_USAXS()
+
+    if terms.preUSAXStune.use_specific_location.get() in (1, "yes"):
+        yield from bps.mv(
+            s_stage.x,
+            terms.preUSAXStune.sx.get(),
+            s_stage.y,
+            terms.preUSAXStune.sy.get(),
+            timeout=MASTER_TIMEOUT,
+        )
+
+    yield from bps.mv(
+        d_stage.x,
+        terms.USAXS.DX0.get(),
+        d_stage.y,
+        terms.USAXS.diode.dy.get(),
+        user_data.time_stamp,
+        str(datetime.datetime.now()),
+        usaxs_slit.v_size,
+        terms.SAXS.usaxs_v_size.get(),
+        usaxs_slit.h_size,
+        terms.SAXS.usaxs_h_size.get(),
+        guard_slit.v_size,
+        terms.SAXS.usaxs_guard_v_size.get(),
+        guard_slit.h_size,
+        terms.SAXS.usaxs_guard_h_size.get(),
+        scaler0.preset_time,
+        0.1,
+        timeout=MASTER_TIMEOUT,
+    )
+    yield from user_data.set_state_plan("pre-USAXS optics tune")
+
+    # when all that is complete, then ...
+    yield from bps.mv(usaxs_shutter, "open", timeout=MASTER_TIMEOUT)
+
+    # TODO: install suspender using usaxs_CheckBeamStandard.get()
+
+    tuners = OrderedDict()  # list the axes to tune
+    # 20IDB does not need tuning M stage too often. Leave to manual staff action
+    # tuners[m_stage.r] = tune_mr            # tune M stage to monochromator
+    if not m_stage.isChannelCut:
+        tuners[m_stage.r2p] = tune_m2rp  # make M stage crystals parallel
+    if terms.USAXS.useMSstage.get():
+        # tuners[ms_stage.rp] = tune_msrp    # align MSR stage with M stage
+        pass
+    if terms.USAXS.useSBUSAXS.get():
+        # tuners[as_stage.rp] = tune_asrp
+        #     align ASR stage with MSR stage
+        #     and set ASRP0 value
+        pass
+    tuners[a_stage.r] = tune_ar  # tune A stage to M stage
+    #tuners[a_stage.r2p] = tune_a2rp  # make A stage crystals parallel
+    tuners[a_stage.r] = tune_ar  # tune A stage to M stage
+    tuners[a_stage.r2p] = tune_a2rp  # make A stage crystals parallel
+
+    # now, tune the desired axes, bail out if a tune fails
+    # yield from bps.install_suspender(suspend_BeamInHutch)
+    for axis, tune in tuners.items():
+        yield from bps.mv(usaxs_shutter, "open", timeout=MASTER_TIMEOUT)
+        yield from tune(md=md)
+        if not axis.tuner.tune_ok:
+            logger.warning("!!! tune failed for axis %s !!!", axis.name)
+            # if NOTIFY_ON_BADTUNE:
+            #     email_notices.send(
+            #         f"USAXS tune failed for axis {axis.name}",
+            #         f"USAXS tune failed for axis {axis.name}",
+            #     )
+
+        # If we don't wait, the next tune often fails
+        # intensity stays flat, statistically
+        # We need to wait a short bit to allow EPICS database
+        # to complete processing and report back to us.
+        yield from bps.sleep(1)
+    # tune a2rp one more time as final step, we will see if it is needed...
+    # yield from bps.mv(usaxs_shutter, "open", timeout=MASTER_TIMEOUT)
+    # yield from tune_a2rp(md=md)
+    # if not axis.tuner.tune_ok:
+    #    logger.warning("!!! tune failed for axis %s !!!", "a2rp")
+    # yield from bps.sleep(1)
+    # yield from bps.remove_suspender(suspend_BeamInHutch)
+
+    logger.info("USAXS count time: %s second(s)", terms.USAXS.usaxs_time.get())
+    yield from bps.mv(
+        scaler0.preset_time,
+        terms.USAXS.usaxs_time.get(),
+        user_data.time_stamp,
+        str(datetime.datetime.now()),
+        terms.preUSAXStune.num_scans_last_tune,
+        0,
+        terms.preUSAXStune.run_tune_next,
+        0,
+        terms.preUSAXStune.epoch_last_tune,
+        time.time(),
+        timeout=MASTER_TIMEOUT,
+    )
+    yield from user_data.set_state_plan("pre-USAXS optics tune")
+
+
+@bpp.suspend_decorator(suspend_FE_shutter)
+@bpp.suspend_decorator(suspend_BeamInHutch)
+def allUSAXStune(
+    md: Optional[Dict[str, Any]] = None,
+    RE: Optional[Any] = None,
+    bec: Optional[Any] = None,
+    specwriter: Optional[Any] = None,
+):
+    """
+    Tune mr, ar, a2rp, ar, a2rp USAXS optics.
+
+    Parameters
+    ----------
+    md : Optional[Dict[str, Any]], optional
+        Metadata dictionary, by default None
+    RE : Optional[Any], optional
+        Bluesky RunEngine instance, by default None
+    bec : Optional[Any], optional
+        Bluesky Live Callbacks instance, by default None
+    specwriter : Optional[Any], optional
+        SPEC file writer instance, by default None
+
+    Returns
+    -------
+    Generator[Any, None, None]
+        A generator that yields plan steps
+
+    USAGE:  ``RE(allUSAXStune())``
+    """
+
+    # Get devices from oregistry
+    monochromator = oregistry["monochromator"]
+    mono_shutter = oregistry["mono_shutter"]
+    ccd_shutter = oregistry["ccd_shutter"]
+    terms = oregistry["terms"]
+    s_stage = oregistry["s_stage"]
+    d_stage = oregistry["d_stage"]
+    user_data = oregistry["user_device"]
+    usaxs_slit = oregistry["usaxs_slit"]
+    guard_slit = oregistry["guard_slit"]
+    scaler0 = oregistry["scaler0"]
+    usaxs_shutter = oregistry["usaxs_shutter"]
+    m_stage = oregistry["m_stage"]
+    a_stage = oregistry["a_stage"]
+
+    yield from bps.mv(
+        monochromator.feedback.on,
+        MONO_FEEDBACK_ON,
+        mono_shutter,
+        "open",
+        ccd_shutter,
+        "close",
+        timeout=MASTER_TIMEOUT,
+    )
+    yield from IfRequestedStopBeforeNextScan(
+        oregistry=oregistry
+    )  # stop if user chose to do so.
+
+    yield from mode_USAXS(oregistry=oregistry)
+
+    if terms.preUSAXStune.use_specific_location.get() in (1, "yes"):
+        yield from bps.mv(
+            s_stage.x,
+            terms.preUSAXStune.sx.get(),
+            s_stage.y,
+            terms.preUSAXStune.sy.get(),
+            timeout=MASTER_TIMEOUT,
+        )
+
+    yield from bps.mv(
+        d_stage.x,
+        terms.USAXS.DX0.get(),  # TODO: resolve this database issue
+        d_stage.y,
+        terms.USAXS.diode.dy.get(),
+        user_data.time_stamp,
+        str(datetime.datetime.now()),
+        usaxs_slit.v_size,
+        terms.SAXS.usaxs_v_size.get(),
+        usaxs_slit.h_size,
+        terms.SAXS.usaxs_h_size.get(),
+        guard_slit.v_size,
+        terms.SAXS.usaxs_guard_v_size.get(),
+        guard_slit.h_size,
+        terms.SAXS.usaxs_guard_h_size.get(),
+        scaler0.preset_time,
+        0.1,
+        timeout=MASTER_TIMEOUT,
+    )
+    yield from user_data.set_state_plan("pre-USAXS optics tune")
+
+    # when all that is complete, then ...
+    yield from bps.mv(usaxs_shutter, "open", timeout=MASTER_TIMEOUT)
+
+    # TODO: install suspender using usaxs_CheckBeamStandard.get()
+
+    tuners = OrderedDict()  # list the axes to tune
+    tuners[m_stage.r] = tune_mr  # tune M stage to monochromator
+    if not m_stage.isChannelCut:
+        tuners[m_stage.r2p] = tune_m2rp  # make M stage crystals parallel
+    if terms.USAXS.useMSstage.get():
+        # tuners[ms_stage.rp] = tune_msrp    # align MSR stage with M stage
+        pass
+    if terms.USAXS.useSBUSAXS.get():
+        # tuners[as_stage.rp] = tune_asrp    # align ASR stage with MSR stage
+        pass
+    tuners[a_stage.r] = tune_ar  # tune A stage to M stage
+    tuners[a_stage.r2p] = tune_a2rp  # make A stage crystals parallel
+    tuners[a_stage.r] = tune_ar  # tune A stage to M stage
+    tuners[a_stage.r2p] = tune_a2rp  # make A stage crystals parallel
+
+    # now, tune the desired axes, bail out if a tune fails
+    # yield from bps.install_suspender(suspend_BeamInHutch)
+    for axis, tune in tuners.items():
+        yield from bps.mv(usaxs_shutter, "open", timeout=MASTER_TIMEOUT)
+        yield from tune(md=md, oregistry=oregistry)
+        if not axis.tuner.tune_ok:
+            logger.warning("!!! tune failed for axis %s !!!", axis.name)
+            if NOTIFY_ON_BADTUNE:
+                email_notices.send(
+                    f"USAXS tune failed for axis {axis.name}",
+                    f"USAXS tune failed for axis {axis.name}",
+                )
+
+        # If we don't wait, the next tune often fails
+        # intensity stays flat, statistically
+        # We need to wait a short bit to allow EPICS database
+        # to complete processing and report back to us.
+        yield from bps.sleep(1)
+    # yield from bps.remove_suspender(suspend_BeamInHutch)
+
+    logger.info("USAXS count time: %s second(s)", terms.USAXS.usaxs_time.get())
+    yield from bps.mv(
+        scaler0.preset_time,
+        terms.USAXS.usaxs_time.get(),
+        user_data.time_stamp,
+        str(datetime.datetime.now()),
+        terms.preUSAXStune.num_scans_last_tune,
+        0,
+        terms.preUSAXStune.run_tune_next,
+        0,
+        terms.preUSAXStune.epoch_last_tune,
+        time.time(),
+        timeout=MASTER_TIMEOUT,
+    )
+    yield from user_data.set_state_plan("pre-USAXS optics tune")
+
+
+# - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
+
+
+def preSWAXStune(
+    md: Optional[Dict[str, Any]] = None,
+    RE: Optional[Any] = None,
+    bec: Optional[Any] = None,
+    specwriter: Optional[Any] = None,
+):
+    """
+    Tune the SAXS/WAXS optics in any mode, is safe.
+
+    Parameters
+    ----------
+    md : Optional[Dict[str, Any]], optional
+        Metadata dictionary, by default None
+    RE : Optional[Any], optional
+        Bluesky RunEngine instance, by default None
+    bec : Optional[Any], optional
+        Bluesky Live Callbacks instance, by default None
+    specwriter : Optional[Any], optional
+        SPEC file writer instance, by default None
+
+    Returns
+    -------
+    Generator[Any, None, None]
+        A generator that yields plan steps
+
+    USAGE:  ``RE(preSWAXStune())``
+    """
+    if md is None:
+        md = {}
+    if RE is None:
+        raise ValueError("RunEngine instance must be provided")
+    if bec is None:
+        raise ValueError("Bluesky Live Callbacks instance must be provided")
+    if specwriter is None:
+        raise ValueError("SPEC file writer instance must be provided")
+
+    # Get devices from oregistry
+    monochromator = oregistry["monochromator"]
+    mono_shutter = oregistry["mono_shutter"]
+    ccd_shutter = oregistry["ccd_shutter"]
+    terms = oregistry["terms"]
+    s_stage = oregistry["s_stage"]
+    user_data = oregistry["user_data"]
+    scaler0 = oregistry["scaler0"]
+    usaxs_shutter = oregistry["usaxs_shutter"]
+    m_stage = oregistry["m_stage"]
+    ms_stage = oregistry["ms_stage"]
+    tune_msrp = oregistry["tune_msrp"]
+    tune_mr = oregistry["tune_mr"]
+    tune_m2rp = oregistry["tune_m2rp"]
+
+    yield from bps.mv(
+        monochromator.feedback.on,
+        MONO_FEEDBACK_ON,
+        mono_shutter,
+        "open",
+        ccd_shutter,
+        "close",
+        timeout=MASTER_TIMEOUT,
+    )
+    yield from IfRequestedStopBeforeNextScan(
+        oregistry=oregistry
+    )  # stop if user chose to do so.
+
+    if terms.preUSAXStune.use_specific_location.get() in (1, "yes"):
+        yield from bps.mv(
+            s_stage.x,
+            terms.preUSAXStune.sx.get(),
+            s_stage.y,
+            terms.preUSAXStune.sy.get(),
+            timeout=MASTER_TIMEOUT,
+        )
+
+    yield from bps.mv(
+        user_data.time_stamp,
+        str(datetime.datetime.now()),
+        scaler0.preset_time,
+        0.1,
+        timeout=MASTER_TIMEOUT,
+    )
+    yield from user_data.set_state_plan("pre-SWAXS optics tune")
+
+    # when all that is complete, then ...
+    yield from bps.mv(usaxs_shutter, "open", timeout=MASTER_TIMEOUT)
+
+    # TODO: install suspender using usaxs_CheckBeamStandard.get()
+
+    tuners = OrderedDict()  # list the axes to tune
+    tuners[m_stage.r] = tune_mr  # tune M stage to monochromator
+    if not m_stage.isChannelCut:
+        tuners[m_stage.r2p] = tune_m2rp  # make M stage crystals parallel
+    if terms.USAXS.useMSstage.get():
+        tuners[ms_stage.rp] = tune_msrp  # align MSR stage with M stage
+
+    # now, tune the desired axes, bail out if a tune fails
+    yield from bps.install_suspender(suspend_BeamInHutch)
+    for axis, tune in tuners.items():
+        yield from bps.mv(usaxs_shutter, "open", timeout=MASTER_TIMEOUT)
+        yield from tune(md=md, oregistry=oregistry)
+        if axis.tuner.tune_ok:
+            # If we don't wait, the next tune often fails
+            # intensity stays flat, statistically
+            # We need to wait a short bit to allow EPICS database
+            # to complete processing and report back to us.
+            yield from bps.sleep(1)
+        else:
+            logger.warning("!!! tune failed for axis %s !!!", axis.name)
+            # break
+    yield from bps.remove_suspender(suspend_BeamInHutch)
+
+    logger.info("USAXS count time: %s second(s)", terms.USAXS.usaxs_time.get())
+    yield from bps.mv(
+        scaler0.preset_time,
+        terms.USAXS.usaxs_time.get(),
+        user_data.time_stamp,
+        str(datetime.datetime.now()),
+        terms.preUSAXStune.num_scans_last_tune,
+        0,
+        terms.preUSAXStune.run_tune_next,
+        0,
+        terms.preUSAXStune.epoch_last_tune,
+        time.time(),
+        timeout=MASTER_TIMEOUT,
+    )
+    yield from user_data.set_state_plan("pre-SWAXS optics tune")
